comment

+++ app/Http/Controllers/CartController.php
@@ -13,34 +13,43 @@ class CartController extends Controller
         return view('cart.viewCart', compact('cart'));
     }
 
-    public function addcart(Request $request, Product $product)
-    {
-        $existingCart = Cart::where('user_id', auth()->id())
-            ->where('store_id', $product->store_id)
-            ->where('product_id', $product->id)
-            ->first();
+public function addcart(Request $request, Product $product)
+{
+    $existingCart = Cart::where('user_id', auth()->id())
+        ->where('store_id', $product->store_id)
+        ->where('product_id', $product->id)
+        ->first();
 
-        if ($existingCart) {
-            return redirect()->route('index.cart');
+        if ($existingCart &&$product->limit > $product->basket) {
+            $product->increment('basket');
+            return redirect('/');
         }
 
+    if ($product->limit > $product->basket) {
+        $product->increment('basket');
         $cart = new Cart();
         $cart->user_id = auth()->id();
         $cart->store_id = $product->store_id;
         $cart->product_id = $product->id;
         $cart->save();
 
-        return redirect()->route('index.cart');
+        return redirect('/');
+    }
+    return redirect()->route('index.cart');
+}
+
+public function deletecart(Cart $cart, Product $product)
+{
+    $product = $cart->product;
+
+    if ($product->basket > 0) {
+        $product->decrement('basket');
     }
 
-    public function deletecart(Cart $cart, Product $product)
-    {
-   
-
-
+    if ($product->basket <= 0) {
         $cart->delete();
-
-        return redirect()->route('index.cart');
     }
 
+    return redirect()->route('index.cart');
+}
 }
